+++ core/src/num/unit.rs
@@ -5,10 +5,11 @@ use crate::num::dist::Dist;
 use crate::num::{Base, FormattingStyle};
 use crate::scope::Scope;
 use crate::serialize::{deserialize_bool, deserialize_usize, serialize_bool, serialize_usize};
+use crate::units::{lookup_default_unit, query_unit_static};
 use crate::{ast, ident::Ident};
 use crate::{Attrs, Span, SpanKind};
 use std::borrow::Cow;
-use std::collections::HashMap;
+use std::collections::{HashMap, HashSet};
 use std::ops::Neg;
 use std::sync::Arc;
 use std::{fmt, io};
@@ -727,7 +728,12 @@ impl Value {
 		})
 	}
 
-	pub(crate) fn simplify<I: Interrupt>(self, int: &I) -> Result<Self, FendError> {
+	pub(crate) fn simplify<I: Interrupt>(
+		self,
+		attrs: Attrs,
+		ctx: &mut crate::Context,
+		int: &I,
+	) -> Result<Self, FendError> {
 		if !self.simplifiable {
 			return Ok(self);
 		}
@@ -815,8 +821,7 @@ impl Value {
 
 		// remove units with exponent == 0
 		res_components.retain(|unit_exponent| unit_exponent.exponent != 0.into());
-
-		Ok(Self {
+		let result = Self {
 			value: res_value,
 			unit: Unit {
 				components: res_components,
@@ -825,7 +830,23 @@ impl Value {
 			base: self.base,
 			format: self.format,
 			simplifiable: self.simplifiable,
-		})
+		};
+
+		if result.unit.has_pos_and_neg_base_unit_exponents() {
+			// try and replace unit with a default one, e.g. `kilogram` or `ampere`
+			let (hashmap, _) = result.unit.to_hashmap_and_scale(int)?;
+			let mut base_units = hashmap
+				.into_iter()
+				.map(|(k, v)| v.try_as_i64(int).map(|v| format!("{}^{v}", k.name())))
+				.collect::<Result<Vec<String>, _>>()?;
+			base_units.sort();
+			if let Some(new_unit) = lookup_default_unit(&base_units.join(" ")) {
+				let rhs = query_unit_static(new_unit, attrs, ctx, int)?.expect_num()?;
+				return result.convert_to(rhs, int);
+			}
+		}
+
+		Ok(result)
 	}
 
 	pub(crate) fn unit_equal_to(&self, rhs: &str) -> bool {
@@ -954,6 +975,28 @@ impl Unit {
 			cs.push(UnitExponent::deserialize(read)?);
 		}
 		Ok(Self { components: cs })
+	}
+
+	fn has_pos_and_neg_base_unit_exponents(&self) -> bool {
+		if self.components.len() <= 1 {
+			return false;
+		}
+
+		let mut pos = HashSet::new();
+		let mut neg = HashSet::new();
+		for comp in &self.components {
+			let component_sign = comp.exponent > 0.into();
+			for (base, base_exp) in &comp.unit.base_units {
+				let base_sign = base_exp > &0.into();
+				let combined_sign = component_sign == base_sign; // xnor
+				if combined_sign {
+					pos.insert(base);
+				} else {
+					neg.insert(base);
+				}
+			}
+		}
+		pos.intersection(&neg).next().is_some()
 	}
 
 	pub(crate) fn equal_to(&self, rhs: &str) -> bool {